requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
Pillow==8.0.1
pkg-resources==0.0.0
pycparser==2.20
PyPDF2==1.26.0
python-telegram-bot==13.0
pytz==2020.4
qrcode==6.1
reportlab==3.5.55
six==1.15.0
SQLAlchemy==1.3.20
tornado==6.1
tzlocal==2.1
